festive-ferrets/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
